+++ lib/mplcairo/base.py
@@ -177,14 +177,7 @@ class GraphicsContextRendererCairo(
         self.draw_image(self, l + dx, height - b - h + dy, img)
 
     start_rasterizing = _mplcairo.GraphicsContextRendererCairo.start_filter
-
-    # While we could just write
-    #   stop_rasterizing = partialmethod(stop_filter,
-    #                                    lambda img, dpi: (img, 0, 0))
-    # this crashes inspect.signature on Py3.6
-    # (https://bugs.python.org/issue33009).
-    def stop_rasterizing(self):
-        return self.stop_filter(lambda img, dpi: (img, 0, 0))
+    stop_rasterizing = partialmethod(stop_filter, lambda img, dpi: (img, 0, 0))
 
     # "Undocumented" APIs needed to patch Agg.