+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 next
 ====
 
-- Bumped dependencies to pybind11≥2.6.
+- Bumped dependencies to Python≥3.7, pybind11≥2.6.
 - Improve snapping of codeful paths.
 - Fix failure to increase cairo refcount of font cache entries.
 - Support ``rcParams["pdf.compression"] = 0``.